+++ CotEditor/Sources/ATSTypesetter.swift
@@ -10,7 +10,7 @@
  ------------------------------------------------------------------------------
  
  © 2004-2007 nakamuxu
- © 2014-2016 1024jp
+ © 2014-2017 1024jp
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -60,7 +60,7 @@ final class ATSTypesetter: NSATSTypesetter {
             
             let isLowSurrogate = CFStringIsSurrogateLowCharacter(string.character(at: charIndex)) && CFStringIsSurrogateHighCharacter(string.character(at: charIndex - 1))
             if !isLowSurrogate {
-                return .whitespaceAction  // -> Then, the glyph width can be modified on `boundingBoxForControlGlyphAtIndex:...`.
+                return .whitespaceAction  // -> Then, the glyph width can be modified on `boundingBox(forControlGlyphAt:...)`.
             }
         }
         
@@ -78,13 +78,8 @@ final class ATSTypesetter: NSATSTypesetter {
         }
         
         // make blank space to draw a replacement character in LayoutManager later.
-        guard let textFont = manager.textFont else { return .zero }
-        let invisibleFont = NSFont(name: "Lucida Grande", size: textFont.pointSize) ?? textFont  // use current text font for fallback
-        let replacementGlyph = invisibleFont.glyph(withName: "replacement")  // U+FFFD
-        let replacementGlyphBounding = invisibleFont.boundingRect(forGlyph: replacementGlyph)
-        
         var rect = proposedRect
-        rect.size.width = replacementGlyphBounding.width
+        rect.size.width = manager.replacementGlyphWidth
         
         return rect
     }
@@ -93,18 +88,23 @@ final class ATSTypesetter: NSATSTypesetter {
     /// avoid soft warpping just after an indent
     override func shouldBreakLine(byWordBeforeCharacterAt charIndex: Int) -> Bool {
         
-        if charIndex == 0 { return true }
+        // -> Getting index fails when the code point is a part of surrogate pair.
+        guard
+            charIndex == 0,
+            let string = self.attributedString?.string,
+            let index = String.UTF16Index(encodedOffset: charIndex).samePosition(in: string)
+            else { return true }
         
         // check if the character is the first non-whitespace character after indent
-        guard let string = self.attributedString?.string else { return true }
-        
-        // -> Getting index fails when the code point is a part of surrogate pair.
-        guard let index = String.UTF16Index(encodedOffset: charIndex).samePosition(in: string) else { return true }
-        
-        // -> Don't use `characters` instead of `unicodeScalars` because of a memory leak (2016-09 macOS 10.12).
-        for character in string.substring(to: index).unicodeScalars.reversed() {
-            if character == "\n" { return false }  // the line ended before hitting to any indent characters
-            if character != " " && character != "\t" { return true }  // hit to non-indent character
+        for character in string[..<index].reversed() {
+            switch character {
+            case "\n":  // the line ended before hitting to any indent characters
+                return false
+            case " ", "\t":
+                continue
+            default:  // hit to non-indent character
+                return true
+            }
         }
         
         return false  // didn't hit to line-break (= first line)