+++ CotEditor/Sources/WindowContentViewController.swift
@@ -110,6 +110,27 @@ final class WindowContentViewController: NSSplitViewController, TabViewControlle
     }
     
     
+    /// disable toggling sidebar in the tab overview mode
+    override func validateToolbarItem(_ item: NSToolbarItem) -> Bool {
+        
+        guard let action = item.action else { return false }
+        
+        switch action {
+        case #selector(getInfo), #selector(toggleIncompatibleCharList):
+            if #available(macOS 10.13, *),
+                let window = self.view.window, window.isVisible,  // check visiblity to avoid the window position cascading bug
+                let tabGroup = window.tabGroup {
+                return !tabGroup.isOverviewVisible
+            }
+            return true
+            
+        default: break
+        }
+        
+        return super.validateToolbarItem(item)
+    }
+    
+    
     
     // MARK: Sidebar View Controller Delegate