+++ CHANGELOG.md
@@ -2,7 +2,28 @@
 Change Log
 ==========================
 
-3.2.1 (203)
+3.2.3 (205)
+--------------------------
+
+### Improvements
+
+- Disable toggling sidebar while the tab overview mode on High Sierra.
+- Update “CSS” syntax style:
+    - Fix an issue where keywords were highlighted incorrectly.
+- Some minor improvements and fixes.
+
+
+### Fixes
+
+- Fix an issue where UNIX scripts could fail getting the content of the document.
+- Fix an issue where font change in the preferences pane could be ignored.
+- Fix a potential issue where syntax keywords could be highlighted incorrectly if whitespaces accidentally get into keywords definition.
+- Workaround an issue where word suggestion in the Touch Bar cannot insert a word starts with a symbol correctly.
+- Workaround an issue where the application could crash on document auto-saving.
+
+
+
+3.2.2 (203)
 --------------------------
 
 ### New Features